Initial Implementation of GRU layer
Following the implementation of BatchNorm, use the underlying Flax
implementation for the logic behind the forward pass. This also adds the
use of `jax.lax.scan` for performing the sequential calls on the hidden
state, making this the implementation of the GRU layer as opposed to a
reimplementation of the `GRUCell`.