Change `time_major` to `time_axis`
In preparation for the updates to `flax.jax_utils.scan_in_dim`, this
changes `time_major` to `time_axis`. Currently allowing for only the
specification of a single time dimension via `type hinting` although
this is currently not enforced in runtime. But underneath, it stores
this as a tuple which would allow for its use in `scan_in_dim`.