Adds GRU layer (#48)
* Initial Implementation of GRU layer

Following the implementation of BatchNorm, use the underlying Flax
implementation for the logic behind the forward pass. This also adds the
use of `jax.lax.scan` for performing the sequential calls on the hidden
state, making this the implementation of the GRU layer as opposed to a
reimplementation of the `GRUCell`.

* Add key functionality to GRU implementation

- `return_state`: whether the final state should be returned
- `return_sequences`: whether all the intermediate states should be
  returned
- `go_backwards`: whether the input should be run in reverse order

* Swap order of state and input variables

Swaps the order of the state and input variables as well as the return
order of the final state and sequence of states to be more inline with
that of Keras API.

* Allow for optional passing of initial state

This allows for optional passing an initial state and the specifying of
a function which can intialize the initial state.

* Add `stateful` flag

Adds the `stateful` flag which allows for the last state of the GRU to
be used as the start state for the next batch.

* Adds documentation to GRU

* Change `time_major` to `time_axis`

In preparation for the updates to `flax.jax_utils.scan_in_dim`, this
changes `time_major` to `time_axis`. Currently allowing for only the
specification of a single time dimension via `type hinting` although
this is currently not enforced in runtime. But underneath, it stores
this as a tuple which would allow for its use in `scan_in_dim`.

* Sets default `time_axis=-2`

Changes the default `time_axis` to be -2, i.e. by default the expected
shape of the input should be of the form [..., time, :, :].

* Swap `jax.lax.scan` for `jax_utils.scan_in_dim`

Swaps `jax.lax.scan` for `flax.jax_utils.scan_in_dim` which allows for
one to have multiple time dimensions by specifying `time_axis` to be a
tuple instead of a single int value.

* Revert "Swap `jax.lax.scan` for `jax_utils.scan_in_dim`"

This reverts commit 4a8f15a590c421bac6e9012925849965b5875ea2.

* Fix testcases due to different default timeaxis

* Fix black formatting issues