Bumps `flax` to `0.4.0`
Updates flax to the most recent version. This currently breaks the
current implementation and way in which rng keys of  dropout is being
handled.

Currently have disabled one of the dropout equivalence tests as I am not
fully aware if there is a method of directly affecting the value of
`next_key` within a treex module.